test(ui): cover prompt handoff integration
